pyproject.toml
[tool.poetry]
name = "finallproject"
version = "0.1.0"

description = "app_parking"
authors = ["IVadymMarchenko <topim31@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-cloudinary-storage = "^0.3.0"
cloudinary = "^1.41.0"
imutils = "^0.5.4"
matplotlib = "^3.9.2"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"
paddleocr = "^2.8.1"
paddlepaddle = "^2.6.1"
setuptools = "^73.0.1"
celery = "^5.4.0"
redis = "^5.0.8"
django-redis = "^5.4.0"
django-celery-results = "^2.5.1"
eventlet = "^0.36.1"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
